+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- fix: allow installing forge mods on 1.20.1 neoforge
-- fix(neoforge): handle 1.20.1 versions correctly
-- chore: repackage to io.github.sculk_cli
+- feat: add --request-retries global option to customise the number of times requests should be retried before giving up.